refactor: 로딩중 삭제 및 타입 단언 삭제(#117)
