+++ src/pages/match/create/components/match-card-section.tsx
@@ -1,7 +1,7 @@
 import Card from '@components/card/match-card/card';
 import type { CardProps } from '@components/card/match-card/types/card';
 import { formatToKoreanDate } from '@pages/home/utils/date-format';
-import type { MatchCardData } from '@pages/match/hooks/useMateCreate';
+import type { MatchCardData } from '@pages/match/hooks/use-mate-create';
 
 interface MatchCardSectionProps {
   matchData: MatchCardData;
